+++ apps/client/src/app/pages/portfolio/analysis/analysis-page.component.ts
@@ -321,6 +321,7 @@ export class AnalysisPageComponent implements OnDestroy, OnInit {
           .fetchBenchmarkForUser({
             dataSource,
             symbol,
+            filters: this.userService.getFilters(),
             range: this.user?.settings?.dateRange,
             startDate: this.firstOrderDate
           })