+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added a new static portfolio analysis rule: _Regional Market Cluster Risk_ (Japan)
 - Extended the tags selector component by a `readonly` attribute
 - Extended the tags selector component to support creating custom tags
 - Added global styles to the _Storybook_ setup